Create a new package for common CLI code
In preparation for adding HCV support to nucamoni, this commit splits
the `cli/hiv1b` package into a package of common types and
procedures (`cli/cli`) and a package of procedures specific to
HIV1B. Some further refactoring to reduce code duplication may be
desireable.
